Add basic Snake module with Snake & Coords types.
Coords is a simple type that holds two Integers (representing the X and
Y coordinates). It has a Show instance, which simply shows it in a nice
"(x: 0, y: 0)"-style format.

Snake is a record type that has one field, a List of Coords, which hold
the coordinates of the snake's body. It also has a Show instance, which
simplifies the wrapped Coords' printable form to be more amenable to be
printed in List form (removes the x/y markers and the comma separator).