Update Main module to play extremely barebones snake game.
The main function now generates a random world and feeds it to the snake
function. The snake function prints the world given to it, gets the next
move, feeds it to the updateWorld function (which updates the world with
it), and calls itself with the update world to do it all again.